+++ core/events/bus.go
@@ -147,6 +147,8 @@ const (
 	ReferralProgramStartedEvent
 	ReferralProgramEndedEvent
 	ReferralProgramUpdatedEvent
+	ReferralSetCreatedEvent
+	RefereeJoinedReferralSetEvent
 )
 
 var (
@@ -228,6 +230,8 @@ var (
 		eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_STARTED:         ReferralProgramStartedEvent,
 		eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_ENDED:           ReferralProgramEndedEvent,
 		eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_UPDATED:         ReferralProgramUpdatedEvent,
+		eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_SET_CREATED:             ReferralSetCreatedEvent,
+		eventspb.BusEventType_BUS_EVENT_TYPE_REFEREE_JOINED_REFERRAL_SET:      RefereeJoinedReferralSetEvent,
 		// If adding a type here, please also add it to data-node/broker/convert.go
 	}
 
@@ -300,6 +304,8 @@ var (
 		ReferralProgramStartedEvent:       eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_STARTED,
 		ReferralProgramEndedEvent:         eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_ENDED,
 		ReferralProgramUpdatedEvent:       eventspb.BusEventType_BUS_EVENT_TYPE_REFERRAL_PROGRAM_UPDATED,
+		ReferralSetCreatedEvent:           eventspb.BusEventType_BUS_EVENT_TYPE_REFEREE_JOINED_REFERRAL_SET,
+		RefereeJoinedReferralSetEvent:     eventspb.BusEventType_BUS_EVENT_TYPE_REFEREE_JOINED_REFERRAL_SET,
 		// If adding a type here, please also add it to data-node/broker/convert.go
 	}
 
@@ -373,6 +379,8 @@ var (
 		ReferralProgramStartedEvent:       "ReferralProgramStartedEvent",
 		ReferralProgramEndedEvent:         "ReferralProgramEndedEvent",
 		ReferralProgramUpdatedEvent:       "ReferralProgramUpdatedEvent",
+		ReferralSetCreatedEvent:           "RefferralSetCreatedEvent",
+		RefereeJoinedReferralSetEvent:     "ReffereeJoinReferralSetEvent",
 	}
 )