+++ core/referral/dependencies.go
@@ -14,12 +14,17 @@ package referral
 
 import (
 	"context"
+	"time"
 
 	"code.vegaprotocol.io/vega/core/events"
 	"code.vegaprotocol.io/vega/core/types"
 )
 
-//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/referral EpochEngine,Broker,TeamsEngine
+//go:generate go run github.com/golang/mock/mockgen -destination mocks/mocks.go -package mocks code.vegaprotocol.io/vega/core/referral EpochEngine,Broker,TimeService
+
+type TimeService interface {
+	GetTimeNow() time.Time
+}
 
 // EpochEngine is used to know when to apply the team switches.
 type EpochEngine interface {
@@ -29,11 +34,4 @@ type EpochEngine interface {
 // Broker is used to notify administrative actions on teams and members.
 type Broker interface {
 	Send(event events.Event)
-}
-
-// TeamsEngine is used to retrieve statistics about a team member to compute
-// referral program related data.
-type TeamsEngine interface {
-	IsTeamMember(party types.PartyID) bool
-	NumberOfEpochInTeamForParty(party types.PartyID) uint64
 }