+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - [9045](https://github.com/vegaprotocol/vega/issues/9045) - Snapshot the referral engine.
 - [9136](https://github.com/vegaprotocol/vega/issues/9136) - Referral engine returns reward factor for a given party
 - [9076](https://github.com/vegaprotocol/vega/issues/9076) - Add current funding rate to market data.
+- [9208](https://github.com/vegaprotocol/vega/issues/9208) - Refactor referral set and teams state
 
 ### 🐛 Fixes